# SkillDetectServer/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SkillDetectServer API
  description: |
    Skill 投毒安全检测后端服务接口草案。
    - HMI 前缀 `/api/v1`：面向前端页面（SSO，当前 `security.enabled=false` 不启用）。
    - M2M 前缀 `/api/m2m/v1`：面向流水线（API Key，当前不启用），异步 + 轮询 + SARIF。
    - 健康接口：`/healthz`、`/readyz`、`/api/v1/engine/health`。
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: 单机 docker compose 本地环境
...
security: []
paths:
...
  /api/v1/baselines:
    post:
      tags: [HMI]
      summary: 创建基线（误报抑制，P2）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, content]
              properties:
                name: { type: string }
                content: { type: string }
                format: { type: string, enum: [yaml, json], default: yaml }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaselineResponse'
    get:
      tags: [HMI]
      summary: 基线列表
      responses:
        '200':
          description: 基线列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaselineListResponse'
  /api/v1/baselines/{baselineId}:
    get:
      tags: [HMI]
      summary: 基线详情
      parameters:
        - name: baselineId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 基线详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaselineDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # M2M
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    hmiOidc:
      type: oauth2
      description: HMI SSO/OIDC（当前未启用）
      flows:
        authorizationCode:
          authorizationUrl: https://sso.example.com/oauth/authorize
          tokenUrl: https://sso.example.com/oauth/token
          scopes: { openid: 身份, profile: 基础资料 }
    m2mApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: M2M API Key（当前未启用）
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Envelope:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        requestId: { type: string, example: req_xxxx }
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            code:
              type: integer
              example: 40001
...
    Baseline:
      type: object
      properties:
        baselineId: { type: string }
        name: { type: string }
        format: { type: string }
        version: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    BaselineResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Baseline' }
    BaselineDetail:
      type: object
      properties:
        baselineId: { type: integer, format: int64 }
        name: { type: string }
        content: { type: string }
        format: { type: string }
        version: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    BaselineDetailResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/BaselineDetail' }
    BaselineList:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Baseline' }
    BaselineListResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/BaselineList' }
...